Move check function to WasabiSignerHelpers
